Тип координат изменен с Int на Double

+++ app/src/main/java/com/example/bincardapp/data/api/bin_lookup/models/CountryInfoDto.kt
@@ -4,6 +4,6 @@ import com.google.gson.annotations.SerializedName
 
 data class CountryInfoDto(
     @SerializedName("name") val name: String?,
-    @SerializedName("latitude") val latitude: Int?,
-    @SerializedName("longitude") val longitude: Int?,
+    @SerializedName("latitude") val latitude: Double?,
+    @SerializedName("longitude") val longitude: Double?,
 )
