+++ app/src/main/java/com/example/bincardapp/features/bin_lookup/domain/model/CountryInfoModel.kt
@@ -2,6 +2,6 @@ package com.example.bincardapp.features.bin_lookup.domain.model
 
 data class CountryInfoModel(
     val name: String,
-    val latitude: Int,
-    val longitude: Int,
+    val latitude: Double,
+    val longitude: Double,
 )
